implement ipv6 support

+++ src/ip_addrs.rs
@@ -39,8 +39,8 @@ pub fn calc_netmask_size_with_n_hosts4(n: usize) -> u32 {
 }
 
 /// Calculate the size an IPv6 netmask would need to encompass the number of hosts
-pub fn calc_netmask_size_with_n_hosts6(n: usize) -> u32 {
-    128 - ((n as f64).abs() + 2.0).log2().ceil() as u32
+pub fn calc_netmask_size_with_n_hosts6(n: usize) -> usize {
+    128 - ((n as f64).abs() + 2.0).log2().ceil() as usize
 }
 
 pub fn get_nth_address_in_network4(
